DFCI-GPT4DFCI-Backend/pyproject.toml
[tool.poetry]
name = "dfci-GPT4DFCI-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
azure-cosmos = "^4.5.0"
azure-identity = "^1.14.0"
fastapi = "^0.101.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
openai = "^0.27.8"
mypy = "^1.5.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.284"
black = "^23.7.0"
pytest = "^7.4.0"
httpx = "^0.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
